requirements.txt
django==3.0.8
djangorestframework==3.11.0
psycopg2==2.8.5
pytest==5.4.3
django-configurations==2.2
gunicorn==20.0.4